requirements.txt
flask-0.12.2
flask-wtf-0.14.3
spotipy-2.13.0
mip-1.11.0